feat(parser): move ParseError into a separate module reworking floats' parsing
