+++ javac-rs-parser/src/lib.rs
@@ -1,1 +1,2 @@
+mod error;
 mod parser;
